fix issues

+++ dialog.cpp
@@ -20,6 +20,7 @@ Dialog::Dialog(QWidget *parent) :
     ui->checkBox_BasedProjectOn->setChecked(false);
     on_checkBox_BasedProjectOn_stateChanged(Qt::Unchecked);
     ui->timeEdit_ModifiedTime->setDateTime(QDateTime::currentDateTime());
+    ui->timeEdit_CreatedTime->setDisabled(true);
     CurrentDialog = this;
 }
 
@@ -36,16 +37,15 @@ void Dialog::on_checkBox_BasedProjectOn_stateChanged(int arg1)
         ui->lineEdit_BasedProjectPath->setDisabled(false);
         ui->pushButton_BasedProjectPath->setDisabled(false);
         ui->lineEdit_CreatedAuthor->setDisabled(true);
-        ui->timeEdit_CreatedTime->setDisabled(true);
         ui->lineEdit_ModifiedAuthor->setDisabled(false);
         ui->timeEdit_ModifiedTime->setDisabled(false);
     } else {
         ui->lineEdit_BasedProjectPath->setDisabled(true);
         ui->pushButton_BasedProjectPath->setDisabled(true);
         ui->lineEdit_CreatedAuthor->setDisabled(false);
-        ui->timeEdit_CreatedTime->setDisabled(false);
         ui->lineEdit_ModifiedAuthor->setDisabled(true);
         ui->timeEdit_ModifiedTime->setDisabled(true);
+        ui->timeEdit_CreatedTime->setDateTime(QDateTime::currentDateTime());
     }
 }
 
@@ -130,6 +130,23 @@ void Dialog::on_buttonBox_accepted()
     }
     MainWindow *main = (MainWindow*)parentWidget();
     if (main) {
+        if (!ui->checkBox_BasedProjectOn->isChecked()) {
+            // when create a new project from scratch, load a default $HOME/.ifcfd/casting_default_project.xml
+            // for setting default value
+            QString filename = QDir::homePath()+"/.ifcfd/casting_default_project.xml";
+            QFile file(filename);
+            if (!file.open(QFile::ReadOnly | QFile::Text)) {
+                std::cout << "Error: Can not read file" << qPrintable(filename)
+                             << "; " << qPrintable(file.errorString())
+                                << std::endl;
+            } else {
+                project->clearValue();
+                project->loadConfigFile(filename);
+            }
+            ui->timeEdit_CreatedTime->setDateTime(QDateTime::currentDateTime());
+        } else {
+            ui->timeEdit_ModifiedTime->setDateTime(QDateTime::currentDateTime());
+        }
         project->getInformation()->updateValueFromDialog();
         main->setProject(project);
         wstring projectpath = project->getInformation()->getProjectPath();
@@ -142,33 +159,34 @@ void Dialog::on_buttonBox_accepted()
                 return;
             }
         }
-        // copy stl
-        QString originpath = ui->lineEdit_BasedProjectPath->text()+"/geometries";
-        Mold *mold = project->getMold();
-        QString newpath = QString::fromStdWString(project->getInformation()->getProjectPath()+L"/geometries");
-        QDir dir2(newpath);
-        if (!dir2.exists()) {
-            if (!dir2.mkpath(dir2.absolutePath())) {
-                std::cerr << "Error: Cannnot open dir "
-                             << qPrintable(dir2.absolutePath()) << std::endl;
-                return;
+        if (ui->checkBox_BasedProjectOn->isChecked()) {
+            // copy stl
+            QString originpath = ui->lineEdit_BasedProjectPath->text()+"/geometries";
+            Mold *mold = project->getMold();
+            QString newpath = QString::fromStdWString(project->getInformation()->getProjectPath()+L"/geometries");
+            QDir dir2(newpath);
+            if (!dir2.exists()) {
+                if (!dir2.mkpath(dir2.absolutePath())) {
+                    std::cerr << "Error: Cannnot open dir "
+                                 << qPrintable(dir2.absolutePath()) << std::endl;
+                }
             }
-        }
-        for (vector<MoldConfiguration*>::iterator it = mold->getMolds().begin();
-             it != mold->getMolds().end(); it++) {
-            MoldConfiguration *config = (*it);
-            QString originfile = originpath+"/"+QString::fromStdWString(config->getName());
-            QString newfile = newpath+"/"+QString::fromStdWString(config->getName());
-            if (QFile::exists(newfile)) {
-                QFile::remove(newfile);
+            for (vector<MoldConfiguration*>::iterator it = mold->getMolds().begin();
+                 it != mold->getMolds().end(); it++) {
+                MoldConfiguration *config = (*it);
+                QString originfile = originpath+"/"+QString::fromStdWString(config->getName());
+                QString newfile = newpath+"/"+QString::fromStdWString(config->getName());
+                if (QFile::exists(newfile)) {
+                    QFile::remove(newfile);
+                }
+                QFile::copy(originfile, newfile);
             }
-            QFile::copy(originfile, newfile);
         }
         main->getUi()->action_open->setDisabled(true);
         main->getUi()->action_project->setDisabled(true);
         QString title = QString::fromStdWString(L"墨华高科CFD压铸仿真平台")+" - "
                         +QString::fromStdWString(project->getInformation()->getProjectPath());
-        setWindowTitle(title);
+        main->setWindowTitle(title);
     }
     close();
 }
